Real time data fetcher/requirements.txt
requests
feedparser
pandas
python-dotenv
tweepy
